staging/pkg/mod/golang.org/x/net@v0.0.0-20201110031124-69a78807bb2b/html/testdata/webkit/tests21.dat

#data
<svg><![CDATA[foo]]>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,20): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo"

#data
<math><![CDATA[foo]]>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,21): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       "foo"

#data
<div><![CDATA[foo]]>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,7): expected-dashes-or-doctype
(1,20): expected-closing-tag-but-got-eof
#new-errors
(1:14) cdata-in-html-content
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <!-- [CDATA[foo]] -->

#data
<svg><![CDATA[foo
#errors
(1,5): expected-doctype-but-got-start-tag
(1,17): expected-closing-tag-but-got-eof
#new-errors
(1:18) eof-in-cdata
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo"

#data
<svg><![CDATA[foo
#errors
(1,5): expected-doctype-but-got-start-tag
(1,17): expected-closing-tag-but-got-eof
#new-errors
(1:18) eof-in-cdata
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo"

#data
<svg><![CDATA[
#errors
(1,5): expected-doctype-but-got-start-tag
(1,14): expected-closing-tag-but-got-eof
#new-errors
(1:15) eof-in-cdata
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<svg><![CDATA[]]>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,17): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<svg><![CDATA[]] >]]>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,21): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]] >"

#data
<svg><![CDATA[]] >]]>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,21): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]] >"

#data
<svg><![CDATA[]]
#errors
(1,5): expected-doctype-but-got-start-tag
(1,16): expected-closing-tag-but-got-eof
#new-errors
(1:17) eof-in-cdata
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]]"

#data
<svg><![CDATA[]
#errors
(1,5): expected-doctype-but-got-start-tag
(1,15): expected-closing-tag-but-got-eof
#new-errors
(1:16) eof-in-cdata
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]"

#data
<svg><![CDATA[]>a
#errors
(1,5): expected-doctype-but-got-start-tag
(1,17): expected-closing-tag-but-got-eof
#new-errors
(1:18) eof-in-cdata
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]>a"

#data
<!DOCTYPE html><svg><![CDATA[foo]]]>
#errors
(1,36): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo]"

#data
<!DOCTYPE html><svg><![CDATA[foo]]]]>
#errors
(1,37): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo]]"

#data
<!DOCTYPE html><svg><![CDATA[foo]]]]]>
#errors
(1,38): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo]]]"

#data
<svg><foreignObject><div><![CDATA[foo]]>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,27): expected-dashes-or-doctype
(1,40): expected-closing-tag-but-got-eof
#new-errors
(1:34) cdata-in-html-content
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <div>
|           <!-- [CDATA[foo]] -->

#data
<svg><![CDATA[<svg>]]>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,22): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"

#data
<svg><![CDATA[</svg>a]]>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,24): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "</svg>a"

#data
<svg><![CDATA[<svg>a
#errors
(1,5): expected-doctype-but-got-start-tag
(1,20): expected-closing-tag-but-got-eof
#new-errors
(1:21) eof-in-cdata
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>a"

#data
<svg><![CDATA[</svg>a
#errors
(1,5): expected-doctype-but-got-start-tag
(1,21): expected-closing-tag-but-got-eof
#new-errors
(1:22) eof-in-cdata
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "</svg>a"

#data
<svg><![CDATA[<svg>]]><path>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,28): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"
|       <svg path>

#data
<svg><![CDATA[<svg>]]></path>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,29): unexpected-end-tag
(1,29): unexpected-end-tag
(1,29): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"

#data
<svg><![CDATA[<svg>]]><!--path-->
#errors
(1,5): expected-doctype-but-got-start-tag
(1,33): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"
|       <!-- path -->

#data
<svg><![CDATA[<svg>]]>path
#errors
(1,5): expected-doctype-but-got-start-tag
(1,26): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>path"

#data
<svg><![CDATA[<!--svg-->]]>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,27): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<!--svg-->"
